Make the feed iterator able to tolerate a missing package. (#7)
* Make the feed iterator able to tolerate a missing package.

Missing packages are not common, but can happen. This doesn't means that we are loosing feed items, it just means that no package was created for one minute. Also make some improvements to the file feed examples.

* Fix decoding error in Python 3.5.

In Python 3.6 json.loads() started to accept binary input, but in Python 3.5 you need to pass a string.